Only run PSR-18 on > 7.1

+++ .travis.sh
@@ -103,15 +103,19 @@ http_factory_guzzle_test)
 	php -d extension=modules/psr.so ./http-factory-guzzle/vendor/bin/phpunit -c http-factory-guzzle/
 	;;
 guzzle_psr18_adapter_init)
-	rm -rf guzzle-psr18-adapter
-	git clone -b ${HTTP_GUZZLE_PSR18_ADAPTER_VERSION:-master} https://github.com/ricardofiorani/guzzle-psr18-adapter.git
-	cd guzzle-psr18-adapter
-	composer install $DEFAULT_COMPOSER_FLAGS
-	rm -rf vendor/psr
-	cd ..
+	if [ "$TRAVIS_PHP_VERSION" != "7.0" ]; then
+		rm -rf guzzle-psr18-adapter
+		git clone -b ${HTTP_GUZZLE_PSR18_ADAPTER_VERSION:-master} https://github.com/ricardofiorani/guzzle-psr18-adapter.git
+		cd guzzle-psr18-adapter
+		composer install $DEFAULT_COMPOSER_FLAGS
+		rm -rf vendor/psr
+		cd ..
+	fi
 	;;
 guzzle_psr18_adapter_test)
-	php -d extension=modules/psr.so ./guzzle-psr18-adapter/vendor/bin/phpunit -c guzzle-psr18-adapter/
+	if [ "$TRAVIS_PHP_VERSION" != "7.0" ]; then
+		php -d extension=modules/psr.so ./guzzle-psr18-adapter/vendor/bin/phpunit -c guzzle-psr18-adapter/
+	fi
 	;;
 esac
